+++ src/com/intellij/idea/plugin/hybris/system/bean/psi/reference/OccBSBeanPropertyReference.kt
@@ -19,8 +19,10 @@
 package com.intellij.idea.plugin.hybris.system.bean.psi.reference
 
 import com.intellij.codeInsight.highlighting.HighlightedReference
+import com.intellij.idea.plugin.hybris.psi.util.PsiUtils
 import com.intellij.idea.plugin.hybris.system.bean.codeInsight.completion.BSCompletionService
 import com.intellij.idea.plugin.hybris.system.bean.meta.model.BSGlobalMetaBean
+import com.intellij.idea.plugin.hybris.system.bean.psi.reference.result.BeanPropertyResolveResult
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiPolyVariantReference
@@ -43,6 +45,7 @@ class OccBSBeanPropertyReference(
         return meta.allProperties[propertyName]
             ?.let { BeanPropertyResolveResult(it) }
             ?.let { arrayOf(it) }
+            ?.let { PsiUtils.getValidResults(it) }
             ?: emptyArray()
     }
 }